Initial replacement of TSFetchUrl
We can't use TSFetchUrl() because we need to read and write to the
origin server and we also need to implement HTTP methods other than
GET. This is the initial switch from TSFetchUrl() to the lower level
TSHttpConnect(). It kinda works, but needs some simplification and
robustness (proper locking and refcount and error handling).